backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
qdrant-client
openai
